worker screen- comments

+++ app/src/main/java/com/example/switchingshifts/WorkerScreen.java
@@ -1,7 +1,9 @@
 package com.example.switchingshifts;
 
-import android.app.Notification;
+import android.Manifest;
 import android.content.Intent;
+import android.content.pm.PackageManager;
+import android.net.Uri;
 import android.os.Bundle;
 import android.text.TextUtils;
 import android.view.Menu;
@@ -10,16 +12,15 @@ import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.Button;
+import android.widget.ImageView;
 import android.widget.Spinner;
 import android.widget.TextView;
 import android.widget.Toast;
-
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
-import androidx.core.app.NotificationCompat;
-import androidx.core.app.NotificationManagerCompat;
-
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
@@ -28,10 +29,10 @@ import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.QuerySnapshot;
-
-import java.io.Serializable;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.Calendar;
+import java.util.Collections;
 import java.util.Date;
 import java.util.List;
 import java.util.Stack;
@@ -44,29 +45,31 @@ import backend.Vetrex;
 
 
 /*The worker main screen */
-public class WorkerScreen extends AppCompatActivity implements Serializable {
+public class WorkerScreen extends AppCompatActivity {
     private FirebaseAuth firebase_auth;
     private FirebaseFirestore db;
+    private static final int REQUEST_CALL=1;
     private String user_id, worker_role, shift_reg_selcted, shift_wanted_selcted, shift_reg_id, shift_wanted_id;
+    private String request_id, current_id_user, current_id_shift_reg, current_id_shift_wanted, next_id_user, current_date, phone_number;
     private List<String> shifts_reg = new ArrayList<>();
     private List<String> id_shifts_reg = new ArrayList<>();
     private List<String> shifts_wanted = new ArrayList<>();
     private List<String> id_shifts_wanted = new ArrayList<>();
+    private List<String> workers_id = new ArrayList();
+    private List<String> shifts_to_delete = new ArrayList();
     private Spinner s_shift_reg, s_shift_wanted;
     private ArrayAdapter<CharSequence> adapter_shift_reg, adapter_shift_wanted;
-    private String request_id, current_id_user, current_id_shift_reg, current_id_shift_wanted, next_id_user;
+
     private Graph graph;
     private DFS dfs;
     private Vetrex v_worker_id, v_wanted_shift, v_reg_shift;
     private Shift new_shift;
-    private int num_of_requests;
+    private int  num_of_requests;
     private Button ok_button;
     private Request request;
     private Stack<Vetrex> path;
-    private List<String> workers_id = new ArrayList();
-    private List<String> shifts_to_delete = new ArrayList();
     SimpleDateFormat sfd = new SimpleDateFormat("dd-MM-yyyy");
-    private NotificationManagerCompat notificationManager;
+    private Calendar calendar;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -75,16 +78,20 @@ public class WorkerScreen extends AppCompatActivity implements Serializable {
         Toolbar toolbar = findViewById(R.id.app_bar);
         setSupportActionBar(toolbar);
 
-        notificationManager = NotificationManagerCompat.from(this);
         graph = new Graph();
+
 
         /* Initialize Firebase Auth  and firestore*/
         firebase_auth = FirebaseAuth.getInstance();
         db = FirebaseFirestore.getInstance();
         user_id = firebase_auth.getCurrentUser().getUid();
 
+        calendar = Calendar.getInstance();
+        current_date = sfd.format(calendar.getTime());
         final TextView textViewToChange = findViewById(R.id.Worker_Screen_title);
 
+
+        // add 'hello' +name of worker
         final DocumentReference documentReference = db.collection("workers").document(user_id);
         documentReference.get().addOnSuccessListener(new OnSuccessListener<DocumentSnapshot>() {
             @Override
@@ -94,6 +101,37 @@ public class WorkerScreen extends AppCompatActivity implements Serializable {
                         "Hello " + documentSnapshot.get("first_name"));
             }
         });
+
+        //call to manager button
+
+        db.collection("workers").get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
+            @Override
+            public void onComplete(@NonNull Task<QuerySnapshot> task) {
+                if (task.isSuccessful()) {
+                    for (DocumentSnapshot doc : task.getResult()) {
+                        db.collection("workers").document(doc.getId())
+                                .get().addOnSuccessListener(new OnSuccessListener<DocumentSnapshot>() {
+                            @Override
+                            public void onSuccess(DocumentSnapshot documentSnapshot) {
+                                if (documentSnapshot.getString("role").equals("Manager"))
+                                    phone_number=documentSnapshot.getString("phone_number");
+                            }
+                        });
+                    }
+                }
+            }
+        });
+        ImageView imageCall = findViewById(R.id.image_call);
+
+        imageCall.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                makePhoneCall();
+            }
+        });
+
+
+     //add worker shifts to list
 
         db.collection("workers").document(user_id).collection("shifts").get()
                 .addOnSuccessListener(new OnSuccessListener<QuerySnapshot>() {
@@ -105,16 +143,14 @@ public class WorkerScreen extends AppCompatActivity implements Serializable {
                             shifts_reg.add("");
                             List<DocumentSnapshot> list = queryDocumentSnapshots.getDocuments();
                             for (DocumentSnapshot d : list) {
-                                if(d.get("delete").equals(true)){
+                                Date shift_date = d.getDate("date");
+                                if(shift_date.after(calendar.getTime()) && d.get("delete").equals(false)){
+                                    id_shifts_reg.add(d.getId());
+                                    shifts_reg.add(sfd.format(shift_date) + "  " + d.getString("type"));
+                                }
+                                else if(d.get("delete").equals(true) && !sfd.format(shift_date).equals(current_date)){
                                     db.collection("workers").document(user_id).collection("shifts")
                                             .document(d.getId()).delete();
-                                }
-                                else {
-                                    id_shifts_reg.add(d.getId());
-                                    Date shift_date = d.getDate("date");
-                                    shifts_reg.add(sfd.format(shift_date) + "  " + d.getString("type"));
-                                    // shifts_reg.add(d.getId());
-                                    //add date condition
                                 }
                             }
                         }
@@ -126,53 +162,53 @@ public class WorkerScreen extends AppCompatActivity implements Serializable {
         adapter_shift_reg.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
         s_shift_reg.setAdapter(adapter_shift_reg);
 
+
+        //add other workers shifts from the same role to list
         db.collection("workers").get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
-                    @Override
-                    public void onComplete(@NonNull Task<QuerySnapshot> task) {
-                        shifts_wanted.clear();
-                        id_shifts_wanted.clear();
-                        shifts_wanted.add("");
-                        if (task.isSuccessful()) {
-                            for (DocumentSnapshot doc : task.getResult()) {
-                                if(!doc.getId().equals(user_id)) {
-                                    final String name = doc.getString("first_name");
-                                    final String id = doc.getId();
-                                db.collection("workers").document(doc.getId()).collection("shifts").get()
-                                        .addOnSuccessListener(new OnSuccessListener<QuerySnapshot>() {
-                                            //add Equal to role
-                                            @Override
-                                            public void onSuccess(QuerySnapshot queryDocumentSnapshots) {
-                                                if (!queryDocumentSnapshots.isEmpty()) {
-                                                    List<DocumentSnapshot> list = queryDocumentSnapshots.getDocuments();
-                                                    for (DocumentSnapshot d : list) {
-                                                        if (!shifts_reg.contains(d.getId())) {
-                                                            if(d.getString("role").equals(worker_role)){
-                                                                if(d.get("delete").equals(true)){
-                                                                    db.collection("workers").document(id).collection("shifts")
-                                                                            .document(d.getId()).delete();
-                                                                }else{
-                                                                    id_shifts_wanted.add(d.getId());
-                                                                    Date shift_date = d.getDate("date");
-                                                                    // shifts_wanted.add(d.getId());
-                                                                    shifts_wanted.add(sfd.format(shift_date) + "  " + d.getString("type") + " -" + name);
-                                                                    //add date condition
-                                                                }
-
+            @Override
+            public void onComplete(@NonNull Task<QuerySnapshot> task) {
+                shifts_wanted.clear();
+                id_shifts_wanted.clear();
+                shifts_wanted.add("");
+                if (task.isSuccessful()) {
+                    for (DocumentSnapshot doc : task.getResult()) {
+                        if(!doc.getId().equals(user_id)) {
+                            final String name = doc.getString("first_name");
+                            final String id = doc.getId();
+                            db.collection("workers").document(doc.getId()).collection("shifts").get()
+                                    .addOnSuccessListener(new OnSuccessListener<QuerySnapshot>() {
+                                        //add Equal to role
+                                        @Override
+                                        public void onSuccess(QuerySnapshot queryDocumentSnapshots) {
+                                            if (!queryDocumentSnapshots.isEmpty()) {
+                                                List<DocumentSnapshot> list = queryDocumentSnapshots.getDocuments();
+                                                for (DocumentSnapshot d : list) {
+                                                    if (!shifts_reg.contains(d.getId())) {
+                                                        if(d.getString("role").equals(worker_role)){
+                                                            Date shift_date = d.getDate("date");
+                                                            if(shift_date.after(calendar.getTime()) && d.get("delete").equals(false)){
+                                                                id_shifts_wanted.add(d.getId());
+                                                                shifts_wanted.add(sfd.format(shift_date) + "  " + d.getString("type") + " -" + name);
                                                             }
-
+                                                            else if(d.get("delete").equals(true) && !sfd.format(shift_date).equals(current_date)) {
+                                                                db.collection("workers").document(id).collection("shifts")
+                                                                        .document(d.getId()).delete();
+                                                            }
                                                         }
+
                                                     }
-
                                                 }
+
                                             }
+                                        }
 
-                                        });
-                                }
-                            }
+                                    });
                         }
-
                     }
-                });
+                }
+
+            }
+        });
 
 
         shifts_wanted.add("");
@@ -240,6 +276,7 @@ public class WorkerScreen extends AppCompatActivity implements Serializable {
                     ((TextView) s_shift_wanted.getSelectedView()).setError("חובה למלא שדה זה");
                     flag = true;
                 }
+                //get thr request
                 if (!flag) {
                     request = new Request(shift_reg_id, shift_wanted_id, user_id);
                     request_id = shift_reg_id + "_" + shift_wanted_id;
@@ -260,6 +297,7 @@ public class WorkerScreen extends AppCompatActivity implements Serializable {
 
 
     }
+    //get al the requests from data
 
     private void read_requests_from_data() {
 
@@ -298,6 +336,8 @@ public class WorkerScreen extends AppCompatActivity implements Serializable {
                 );
     }
 
+    //find a cycle at the graph
+
     public void start_dfs() {
         dfs = new DFS(graph);
         boolean has_cycle = true;
@@ -328,7 +368,7 @@ public class WorkerScreen extends AppCompatActivity implements Serializable {
                     workers_id.add(current_id_user);
                     String current_user_request = current_id_shift_reg + "_" + current_id_shift_wanted;
 
-                    switch_shift(next_id_user, current_id_user, current_id_shift_wanted, current_id_shift_reg);
+                    switch_shift(next_id_user, current_id_user, current_id_shift_wanted);
                     db.collection("workers").document(current_id_user).collection("shifts").document(current_id_shift_reg)
                             .update("delete", true);
 
@@ -346,19 +386,8 @@ public class WorkerScreen extends AppCompatActivity implements Serializable {
             }
         }
     }
-
-
-
-//    delete_shifts(shifts_to_delete, workers_id);
-    public void delete_shifts(List<String> shifts_to_delete, List<String> workers_id){
-        for(int i = 0; i < shifts_to_delete.size(); i++) {
-            db.collection("workers").document(workers_id.get(i)).collection("shifts").document(shifts_to_delete.get(i)).delete();
-        }
-        shifts_to_delete.clear();
-        workers_id.clear();
-    }
-
-    public void switch_shift(String next_id_user, final String current_id_user, final String current_id_shift_wanted, String current_id_shift_reg){
+// switch shifts
+    public void switch_shift(String next_id_user, final String current_id_user, final String current_id_shift_wanted){
 
         db.collection("workers").document(next_id_user).collection("shifts").document(current_id_shift_wanted).get()
                 .addOnSuccessListener(new OnSuccessListener<DocumentSnapshot>() {
@@ -370,21 +399,33 @@ public class WorkerScreen extends AppCompatActivity implements Serializable {
                             @Override
                             public void onSuccess(Void aVoid) {
                                 Toast.makeText(getBaseContext(), " התבצע חילוף", Toast.LENGTH_LONG).show();
-//                                send_notification();
                             }
                         });
                     }
                 });
     }
-    public void send_notification(){
-        Notification notification = new NotificationCompat.Builder(this, NotificationHelper.channel1_id)
-                .setSmallIcon(R.drawable.ic_message)
-                .setContentTitle("יש לך הודעה חדשה")
-                .setContentText("נמצאה לך משמרת להחלפה")
-                .setPriority(NotificationCompat.PRIORITY_HIGH)
-                .setCategory(NotificationCompat.CATEGORY_MESSAGE)
-                .build();
-        notificationManager.notify(1, notification);
+
+    private void makePhoneCall() {
+
+        if (ContextCompat.checkSelfPermission(WorkerScreen.this,
+                Manifest.permission.CALL_PHONE) != PackageManager.PERMISSION_GRANTED) {
+            ActivityCompat.requestPermissions(WorkerScreen.this,
+                    new String[]{Manifest.permission.CALL_PHONE}, REQUEST_CALL);
+        } else {
+            String dial = "tel:" + phone_number;
+            startActivity(new Intent(Intent.ACTION_CALL, Uri.parse(dial)));
+        }
+    }
+
+    @Override
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
+        if (requestCode == REQUEST_CALL) {
+            if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+                makePhoneCall();
+            } else {
+                Toast.makeText(this, "Permission DENIED", Toast.LENGTH_SHORT).show();
+            }
+        }
     }
 
     public boolean onCreateOptionsMenu(Menu menu) {
@@ -398,6 +439,8 @@ public class WorkerScreen extends AppCompatActivity implements Serializable {
         if (id == R.id.my_shift) {
             Intent intent = new Intent(WorkerScreen.this, WorkerShifts.class);
             String shifts_to_show = "";
+
+            Collections.sort(shifts_reg);
 
             for(String s: shifts_reg) {
                 shifts_to_show = shifts_to_show + s + "\n";
